Split up the main `Temperature` class into small classes
Split up the `Temperature` class into four smaller classes to redure the hotspot being in one class. It is now split into two.
