+++ src/Converter.Temperature/Kelvin.cs
@@ -1,0 +1,60 @@
+﻿namespace Converter.Temperature
+{
+    internal static class Kelvin
+    {
+        /// <summary>
+        /// The kelvin to celsius conversion.
+        /// </summary>
+        /// <param name="firstTemp"> The temperature to convert. </param>
+        /// <returns>
+        /// The converted temperature.
+        /// </returns>
+        public static double KelvinToCelsius(double firstTemp)
+        {
+            var celsiusTemp = firstTemp - 273.15;
+            return celsiusTemp;
+        }
+
+        /// <summary>
+        /// The kelvin to fahrenheit conversion.
+        /// </summary>
+        /// <param name="firstTemp"> The temperature to convert. </param>
+        /// <exception cref="T:System.ArgumentOutOfRangeException"> If calculated value is beyond the limits of the type. </exception>
+        /// <returns>
+        /// The converted temperature.
+        /// </returns>
+        public static double KelvinToFahrenheit(double firstTemp)
+        {
+            var celsiusTemp = KelvinToCelsius(firstTemp);
+            var fahrenheitTemp = Celsius.CelsiusToFahrenheit(celsiusTemp);
+            return fahrenheitTemp;
+        }
+
+        /// <summary>
+        /// The kelvin to kelvin conversion.
+        /// </summary>
+        /// <param name="firstTemp"> The temperature to convert. </param>
+        /// <returns>
+        /// The converted temperature.
+        /// </returns>
+        public static double KelvinToKelvin(double firstTemp)
+        {
+            return firstTemp;
+        }
+
+        /// <summary>
+        /// The kelvin to gas conversion.
+        /// </summary>
+        /// <param name="firstTemp"> The temperature to convert. </param>
+        /// <exception cref="T:System.ArgumentOutOfRangeException"> Temp too low or too high for gas mark! </exception>
+        /// <returns>
+        /// The converted temperature.
+        /// </returns>
+        public static double KelvinToGas(double firstTemp)
+        {
+            var celsiusTemp = KelvinToCelsius(firstTemp);
+            var gasTemp = Celsius.CelsiusToGas(celsiusTemp);
+            return gasTemp;
+        }
+    }
+}
